rename module; custom release workflow
