--- .github/actions/validator_pypi_publish/add_build_prefix.py
@@ -1,49 +1,0 @@
-import re
-import sys
-import toml
-
-
-def add_package_name_prefix(
-    pyproject_path, pep_503_new_package_name, validator_folder_name
-):
-    # Read the existing pyproject.toml file
-    with open(pyproject_path, "r") as f:
-        content = f.read()
-        parsed_toml = toml.loads(content)
-
-    # get the existing package name
-    existing_name = parsed_toml.get("project", {}).get("name")
-
-    # Update the project name to the new PEP 503-compliant name
-    # The package name would've been converted to PEP 503-compliant anyways
-    # But we use this name since it's been concatenated with the seperator
-    updated_content = re.sub(
-        rf'(^name\s*=\s*")({re.escape(existing_name)})(")',
-        rf"\1{pep_503_new_package_name}\3",
-        content,
-        flags=re.MULTILINE,
-    )
-
-    # Write the modified content back to the pyproject.toml file
-    with open(pyproject_path, "w") as f:
-        f.write(updated_content)
-
-    print(f"Updated project name to '{pep_503_new_package_name}'.")
-    print(f"Added package folder '{validator_folder_name}' in {pyproject_path}")
-
-
-if __name__ == "__main__":
-    if len(sys.argv) < 3:
-        print(
-            "Usage: python script.py <pyproject_path>"
-            " <pep_503_new_package_name> <validator-folder-name>"
-        )
-        sys.exit(1)
-
-    pyproject_path = sys.argv[1]
-    pep_503_new_package_name = sys.argv[2]
-    validator_folder_name = sys.argv[3]
-
-    add_package_name_prefix(
-        pyproject_path, pep_503_new_package_name, validator_folder_name
-    )
