--- client_wrapper/__init__.py
@@ -1,11 +1,0 @@
-from client_wrapper.decorators.llm import tt_webhook_polling_sync
-from client_wrapper.decorators.llm import tt_webhook_polling_sync as simlab_connect
-from client_wrapper.decorators.custom_judge import custom_judge
-from client_wrapper.protocols import JudgeResult
-
-__all__ = [
-    "custom_judge",
-    "tt_webhook_polling_sync",
-    "JudgeResult",
-    "simlab_connect"
-]
