+++ guardrails_simlab_client/protocols.py
@@ -1,6 +1,5 @@
 
 from dataclasses import dataclass
-from enum import Enum
 from typing import Dict, Optional
 
 @dataclass
